Refactor: Move inner MCV classes out as the component has become too large.
