+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/ChatMessagesComponent.java
@@ -37,7 +37,7 @@ import bisq.desktop.main.content.chat.ChatUtil;
 import bisq.desktop.main.content.components.ChatMentionPopupMenu;
 import bisq.desktop.main.content.components.CitationBlock;
 import bisq.desktop.main.content.components.UserProfileSelection;
-import bisq.desktop.main.content.components.chat_messages.list_view.ChatMessagesListView;
+import bisq.desktop.main.content.components.chat_messages.list_view.ChatMessagesListController;
 import bisq.i18n.Res;
 import bisq.settings.SettingsService;
 import bisq.user.identity.UserIdentity;
@@ -94,15 +94,15 @@ public class ChatMessagesComponent {
     }
 
     public void setSearchPredicate(Predicate<? super ChatMessageListItem<? extends ChatMessage, ? extends ChatChannel<? extends ChatMessage>>> predicate) {
-        controller.chatMessagesListView.setSearchPredicate(predicate);
+        controller.chatMessagesListController.setSearchPredicate(predicate);
     }
 
     public void setBisqEasyOfferDirectionOrOwnerFilterPredicate(Predicate<? super ChatMessageListItem<? extends ChatMessage, ? extends ChatChannel<? extends ChatMessage>>> predicate) {
-        controller.chatMessagesListView.setBisqEasyOfferDirectionOrOwnerFilterPredicate(predicate);
+        controller.chatMessagesListController.setBisqEasyOfferDirectionOrOwnerFilterPredicate(predicate);
     }
 
     public void setBisqEasyPeerReputationFilterPredicate(Predicate<? super ChatMessageListItem<? extends ChatMessage, ? extends ChatChannel<? extends ChatMessage>>> predicate) {
-        controller.chatMessagesListView.setBisqEasyPeerReputationFilterPredicate(predicate);
+        controller.chatMessagesListController.setBisqEasyPeerReputationFilterPredicate(predicate);
     }
 
     public void resetSelectedChatMessage() {
@@ -114,7 +114,7 @@ public class ChatMessagesComponent {
     }
 
     public void refreshMessages() {
-        controller.chatMessagesListView.refreshMessages();
+        controller.chatMessagesListController.refreshMessages();
     }
 
     public void enableChatDialog(boolean isEnabled) {
@@ -128,7 +128,7 @@ public class ChatMessagesComponent {
         private final Consumer<UserProfile> openUserProfileSidebarHandler;
         private final UserIdentityService userIdentityService;
         private final CitationBlock citationBlock;
-        private final ChatMessagesListView chatMessagesListView;
+        private final ChatMessagesListController chatMessagesListController;
         private final UserProfileService userProfileService;
         private final SettingsService settingsService;
         private final ChatService chatService;
@@ -149,7 +149,7 @@ public class ChatMessagesComponent {
 
             UserProfileSelection userProfileSelection = new UserProfileSelection(serviceProvider);
 
-            chatMessagesListView = new ChatMessagesListView(serviceProvider,
+            chatMessagesListController = new ChatMessagesListController(serviceProvider,
                     this::mentionUserHandler,
                     this::showChatUserDetailsHandler,
                     this::replyHandler,
@@ -157,7 +157,7 @@ public class ChatMessagesComponent {
 
             model = new Model(chatChannelDomain, chatService);
             view = new View(model, this,
-                    chatMessagesListView.getRoot(),
+                    chatMessagesListController.getView().getRoot(),
                     citationBlock.getRoot(),
                     userProfileSelection);
         }
